enterprise-rag-automation/requirements.txt
flask==2.3.3
google-cloud-bigquery==3.12.0
google-cloud-storage==2.13.0
google-auth==2.22.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
werkzeug==2.3.7
